# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger NewsApi",
        "description": "News Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "operationId": "b8e438773a13bec344011b140c909411",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Yarik"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "yarik@yarik"
                                            },
                                            "role": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-21T12:58:45.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-21T12:58:45.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "users empty"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store a new user",
                "operationId": "c81166c37bc4122ee079a75d0b7c2f09",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Yarik"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "yarik@yarik"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1234qwe"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "1234qwe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Yarik"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "yarik@yarik"
                                                },
                                                "role": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/user/{user}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user by id",
                "operationId": "061657bcbd30151e573bb4985f089707",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Yarik"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "yarik@yarik"
                                        },
                                        "role": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-21T12:58:45.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-21T12:58:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found"
                    },
                    "400": {
                        "description": "error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user by id",
                "operationId": "24fa8042770188d4202bfc8ed971f3fd",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Yarik"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "yarik@yarik"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1234qwe"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "1234qwe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Yarik"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "yarik@yarik"
                                        },
                                        "role": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-21T12:58:45.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-21T12:58:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user by id",
                "operationId": "2a07d05fd5221f56f9e81fe1fac00ae8",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "user ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Logs user into system",
                "operationId": "7457e30c8351093fa2d7df823e62298e",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "yarik@yarik"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "qwerty123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Yarik"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "yarik@yarik"
                                                },
                                                "role": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-21T12:58:45.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-21T12:58:45.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "News deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "incorrect login details"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register user in system",
                "operationId": "6a2220445b080123033e9f77b99033b7",
                "requestBody": {
                    "description": "inputs for regisrer",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Yaroslav"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "yarik@yarik"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "qwerty123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "qwerty123"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object",
                                            "example": {
                                                "name": "Yaroslav",
                                                "email": "yarik@yarik",
                                                "role": 0
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username/password supplied"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Logout"
                ],
                "summary": "Logout user from system",
                "operationId": "ad0abc630daed9d1681c9c878df886a5",
                "responses": {
                    "200": {
                        "description": "Logout"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}